% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bspline_basis}
\alias{bspline_basis}
\title{Derive the B-spline basis function value at a value x}
\arguments{
\item{j}{A \code{integer} referring to the knot interval where \code{x} is to be evaluated.}

\item{m}{A \code{integer} signifying the degree of the B-spline basis.}

\item{knots}{A \code{numeric} vector containing the knots for the B-spline basis.}

\item{x}{A \code{double} corresponding to the value which we seek the B-spline function value.}
}
\value{
A \code{double} that is the function value of the B-spline function evaluated at \code{x} for the \code{j}-th knot given the B-spline basis degree \code{m}.
}
\description{
Computes the B-spline function value evaluated at \code{x} for the \code{j}-th knot, when the degree of the B-spline basis is \code{m}.
}
\details{
Used to recursively build the B-spline basis matrix B.
}
\examples{
\dontrun{
  print(1)
}
}
