% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bspline_basis_matrix}
\alias{bspline_basis_matrix}
\title{Derive the B-spline basis matrix given values of x and the degree m}
\arguments{
\item{x}{A \code{numeric} vector containing the predictior values.}

\item{knots}{A \code{numeric} vector containing the knots for the B-spline basis.}

\item{m}{An \code{integer} that signify the desired degree of the B-spline basis. Default is 3 (cubic).}
}
\value{
The constructed B-spline basis \code{matrix}.
}
\description{
Computes the B-spline basis matrix given the \code{knots} and the desired degree, \code{m}.
}
\details{
Recursively build the B-spline basis matrix B. The construction of B uses the Cox-de Boor recursion formula.
}
\examples{
\dontrun{
  print(1)
}
}
