% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_cov_beta}
\alias{calculate_cov_beta}
\title{Calculates the matrix part, \eqn{L} of the covariance matrix \eqn{\hat{\sigma}^2 \cdot L}}
\arguments{
\item{BTB}{A \code{matrix}. The inner matrix product \eqn{B^{T}WB}.
Alternatively, The inner matrix product \eqn{B^{T}WWB}.
Must have dimensions \eqn{(n+4) \times (n+4)}.}

\item{Q}{A \code{matrix}. The matrix \eqn{Q = (B^{T}WB + \lambda \Omega)^{-1}}.
Must have dimensions \eqn{(n+4) \times (n+4)}.}
}
\value{
A \eqn{(n+4) \times (n+4)} \code{matrix}, which we denote \eqn{L}.
}
\description{
Calculates the matrix part, \eqn{L}, of the covariance matrix \eqn{\hat{\sigma}^2 \cdot L}
}
\details{
Calculates \eqn{L} in the covariance matrix of the smoothing spline coefficients

\eqn{\mathrm{Cov}[\hat{\beta}] = \hat{\sigma}^2 \cdot L}

Where \eqn{L = QB^{T}WBQ}.
}
\examples{
print(1)
}
