% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{calculate_cv_score}
\alias{calculate_cv_score}
\title{Internal Function for Estimating LOO and Generalized Cross-Validation Scores}
\usage{
calculate_cv_score(data, df = 2)
}
\description{
This function is used internally to readily get cross-validation scores.
}
\keyword{internal}
