% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_df}
\alias{calculate_df}
\title{Calculates the effective degrees of freedom \eqn{\mathrm{df}}}
\arguments{
\item{lambda}{A \code{double}. The computitional smoothing parameter.}

\item{weights}{A \code{numeric} vector of weights. Must be of length \eqn{n}.}

\item{B}{A \code{matrix}. The B-spline basis matrix.
Must have dimensions \eqn{(n) \times (n+4)}.}

\item{BTB}{A \code{matrix}. The inner matrix product \eqn{B^{T}WB}.
Must have dimensions \eqn{(n+4) \times (n+4)}.}

\item{Omega}{A \code{matrix}. The penalty matrix.
Must have dimensions \eqn{(n+4) \times (n+4)}.}
}
\value{
A \code{double}. The effective degrees of freedom for the smoothing spline.
}
\description{
Calculates the effective degrees of freedom \eqn{\mathrm{df}}
}
\details{
Calculates the effective degrees of freedom \eqn{\mathrm{df}} by \eqn{\mathrm{tr}[B(B^{T}WB + \lambda \Omega)^{-1}B^{T}]}.
In other words, \eqn{\mathrm{df}} is just the trace of the smoother matrix \eqn{S}.
}
\examples{
print(1)
}
