% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{calculate_interior_knots}
\alias{calculate_interior_knots}
\title{Calculate interior knots}
\usage{
calculate_interior_knots(x)
}
\description{
This function is used internally for calculating interior knots.
}
\keyword{internal}
