% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_pred_var}
\alias{calculate_pred_var}
\title{Calculates the variance of predicted values}
\arguments{
\item{B_new}{A \code{matrix}. The possibly new B-spline basis matrix or its derivative.
Must have dimensions \eqn{(m) \times (n+4)}.}

\item{cov_beta}{A \code{matrix}. The matrix \eqn{\hat{\sigma}^2 \cdot L}.
Must have dimensions \eqn{(n+4) \times (n+4)}. See details}
}
\value{
A \code{numeric} vector of length \eqn{m}. The estimated variances of
        the predicted values based on new observations.
}
\description{
Calculates the variance of predicted values
}
\details{
Calculates estimated variances of smoothing spline predicted values.
In fact we calculate the diagonal elements of

\eqn{\mathrm{Cov}[\hat{\beta}] = \hat{\sigma}^2 \cdot L}

Where \eqn{L = QB^{T}W_{a}BQ}, where
\itemize{
  \item \eqn{Q = (B^{T}WB + \lambda \Omega)^{-1}}
  \item \eqn{W_{a}} is either \eqn{W} or \eqn{WW^{T}}
  \item \eqn{B} is the original \eqn{n \times (n+4)} B-spline matrix.
  \item \eqn{\hat{\sigma}^2 = \mathrm{WRSS} / (n - \mathrm{df})}
}
}
\examples{
print(1)
}
