% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_Q}
\alias{calculate_Q}
\title{Computes the inverse matrix of \eqn{B^{T}WB + \lambda \Omega}}
\arguments{
\item{lambda}{A \code{double}. The computitional smoothing parameter.}

\item{BTB}{A \code{matrix}. The inner matrix product \eqn{B^{T}WB}.
Must have dimensions \eqn{(n+4) \times (n+4)}.}

\item{Omega}{A \code{matrix}. The penalty matrix.
Must have dimensions \eqn{(n+4) \times (n+4)}.}
}
\value{
A \eqn{(n+4) \times (n+4)} \code{matrix}, which we denote \eqn{Q}.
}
\description{
Computes the inverse matrix of \eqn{B^{T}WB + \lambda \Omega}
}
\details{
Calculates \eqn{Q = (B^{T}WB + \lambda \Omega)^{-1}}.
}
\examples{
print(1)
}
