% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_r}
\alias{calculate_r}
\title{Computes trace ratio between BTB and Omega}
\arguments{
\item{BTB}{A \code{matrix}. Must have dimensions \eqn{(n+4) \times (n+4)}, where
\eqn{n} is the total number of unique observations.}

\item{Omega}{A \code{matrix}. Must have dimensions \eqn{(n+4) \times (n+4)}.}
}
\value{
A \code{double}. The trace ratio between BTB and Omega.
}
\description{
Computes trace ratio between BTB and Omega
}
\details{
The trace ratio between BTB and Omega, \eqn{r}, is calculated in the following manner:
\enumerate{
  \item Calculates the trace of \code{BTB} excluding the first \eqn{2} and last \eqn{3} elements. We call this \eqn{T_1}.
  \item Calculates the trace of \code{Omega} excluding the first \eqn{2} and last \eqn{3} elements. We call this \eqn{T_2}.
  \item Calculate and return \eqn{r = T_1 / T_2}
}
\eqn{r} is used in the conversion between \code{sp} and \code{lambda}.
}
\examples{
print(1)
}
