% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{convert_to_data_table}
\alias{convert_to_data_table}
\title{Internal Function for Data Conversion}
\usage{
convert_to_data_table(data)
}
\description{
This function is used internally for converting different data types to
\code{data.table}.
}
\keyword{internal}
