% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crp_cs_data}
\alias{crp_cs_data}
\title{C-Reactive Protein (CRP) Clinical Sample Meausurements}
\format{
## `test_data`
A \code{data.table} with 750 rows and 5 columns:
\describe{
  \item{comparison}{The IVD-MD comparison identifiers.}
  \item{SampleID}{The clinical sample identifiers.}
  \item{ReplicateID}{The replicate measurement identifiers.}
  \item{MP_A}{Measurement results from the first IVD-MD in comparison.}
  \item{MP_B}{Measurement results from the second IVD-MD in comparison.}
}
}
\source{
Noklus
}
\usage{
crp_cs_data
}
\description{
C-Reactive Protein (CRP) Clinical Sample Meausurements
}
\keyword{datasets}
