% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{df_vs_cv}
\alias{df_vs_cv}
\title{Internal Function for Calculating \eqn{\mathrm{df}(c)} Versus Cross-Validation Scores}
\usage{
df_vs_cv(data, df_grid)
}
\description{
This function is used internally to calculate cross-validation scores for a grid of \eqn{\mathrm{df}(c)} values.
}
\keyword{internal}
