% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{df_vs_cv_plot}
\alias{df_vs_cv_plot}
\title{Internal Function for Rendering \eqn{\mathrm{df}(c)} Versus Cross-Validation Score Plots}
\usage{
df_vs_cv_plot(
  plot_data,
  n_breaks_y = 10,
  log_scale = FALSE,
  method_colors = c("black", "red"),
  smudge_linetypes = c("longdash", "solid")
)
}
\description{
This function is used internally to plot \eqn{\mathrm{df}(c)} against cross-validation scores.
}
\keyword{internal}
