% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{df_vs_zeta}
\alias{df_vs_zeta}
\title{Internal Function for Calculating \eqn{\hat{\zeta}} for each \eqn{\mathrm{df}(c)}}
\usage{
df_vs_zeta(data, df_grid)
}
\description{
This function is used internally to calculate \eqn{\hat{\zeta}} for each \eqn{\mathrm{df}(c)}.
}
\keyword{internal}
