% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{df_vs_zeta_plot}
\alias{df_vs_zeta_plot}
\title{Internal Function for Rendering \eqn{\mathrm{df}(c)} Versus \eqn{\hat{\zeta}} Plots}
\usage{
df_vs_zeta_plot(
  plot_data,
  n_breaks_y = 10,
  log_scale = FALSE,
  data_structure_colors = c("black", "red"),
  weighted_linetypes = c("longdash", "solid")
)
}
\description{
This function is used internally to plot \eqn{\mathrm{df}(c)} against \eqn{\hat{\zeta}}.
}
\keyword{internal}
