% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{fast_smoothing_spline}
\alias{fast_smoothing_spline}
\title{Internal Function for Faster Smoothing Spline Fitting}
\usage{
fast_smoothing_spline(
  x,
  y,
  u,
  weights,
  all_knots,
  df = NULL,
  lambda = NULL,
  df_max = 7.5,
  additional_parameters = list(method = "gcv", smudge = 1.2, c_star_min = 0.38,
    c_star_max = 1.5, tol = 0.5, window = 10, iter = 1)
)
}
\description{
This function is used internally for fitting a smoothing spline faster.
}
\keyword{internal}
