% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_predictor_and_response.R
\name{fix_one_predictor_and_response}
\alias{fix_one_predictor_and_response}
\title{Internal Function for shifting roles of MP_A and MP_B if condition is met}
\usage{
fix_one_predictor_and_response(
  data,
  threshold = 0.5,
  single_shift = TRUE,
  force_shift = FALSE
)
}
\description{
This function is used internally within \code{fix_predictor_and_response()}.
}
\keyword{internal}
