% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_predictor_and_response.R
\name{fix_predictor_and_response}
\alias{fix_predictor_and_response}
\title{Change roles of MP_A and MP_B based on repeatability variance ratios}
\usage{
fix_predictor_and_response(data, new_data = NULL, show_fixes = FALSE)
}
\arguments{
\item{data}{A \code{list} or \code{data.table} containing columns \code{Comparison},
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{new_data}{A \code{list} or \code{data.table} containing columns \code{Comparison},
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{show_fixes}{A \code{logical}. If \code{TRUE}, an additonal column is attached to the output.
This column consists of \code{logical} values. If \code{TRUE}, for a particular
\code{comparison} element, the roles of MP_A and MP_B are switched for this IVD-MD comparison.}
}
\value{
A \code{data.table} containing columns \code{Comparison},
        \code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.
        If \code{show_fixes} is \code{TRUE}, the output will also contain
        the column \code{shift}.
}
\description{
The smoothing spline model does not account for measurement error in the predictor.
To reduce endogeneity-related issues, switching roles of MP_A and MP_B may be a
possible solution.

For each comparison, the roles of MP_A and MP_B are switched if the ratio of
repeatability variances of MP_A and MP_B is smaller than 0.5.
}
\examples{
print(1)
}
