% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{get_additional_parameters}
\alias{get_additional_parameters}
\title{Internal Function for Constructing List of Additional Parameters}
\usage{
get_additional_parameters(smoothing_type = "loocv", smudge = 1)
}
\description{
This function is used internally to readily get additional_parameters.
}
\keyword{internal}
