% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_matrices}
\alias{get_matrices}
\title{Reconstruct matrices based on smooth.spline output}
\usage{
get_matrices(auxM)
}
\arguments{
\item{auxM}{A \code{List} containing the \code{auxM} object from the smooth.spline fit.}
}
\value{
A list of two 4-banded \code{matrix} objects and one \code{numeric} vector.
}
\description{
Reconstructs the penalty and B-spline matrices based on flattened data.
}
\details{
Reconstruct matrices based on smooth.spline output

This reconstruction assumes symmetrical 4-banded matrices. End-users should not bother with this function.
}
\examples{
\dontrun{
  print(1)
}
}
