% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{global_precision_estimates2}
\alias{global_precision_estimates2}
\title{Calculate imprecision point estimates of measurements in a given IVD-MD comparison}
\arguments{
\item{data}{\code{list} or \code{data.table}. Must contain the following variables:
\itemize{
  \item \code{SampleID}: Sample identifiers. Must be \code{character}.
  \item \code{ReplicateID}: Replicate measurement identifiers within samples. Must be \code{character}.
  \item \code{MP_A}: Measurement results for the IVD-MD used as response variable.
  \item \code{MP_B}: Measurement results for the IVD-MD used as predictor variable.
}}
}
\value{
A \code{list} of length \code{5} with the following point imprecision estimates:
        \code{Var_A}, \code{Var_B}, \code{CV_A}, \code{CV_B} and \code{lambda}.
        See details for more information on these statistics.
}
\description{
Calculate imprecision point estimates of measurements in a given IVD-MD comparison
}
\details{
Calculates five relevant global imprecision estimates. The term global is
used because these imprecision estimates are based on the whole dataset.
These five imprecision estimates are calculated:
\itemize{
  \item{\code{Var_A: }}{Pooled variance of all sample-variances based on MP_A}
  \item{\code{Var_B: }}{Pooled variance of all sample-variances based on MP_B}
  \item{\code{CV_A: }}{CV estimate based on Var_A and the grand mean of all measurements from MP_A}
  \item{\code{CV_B: }}{CV estimate based on Var_B and the grand mean of all measurements from MP_B}
  \item{\code{lambda: }}{Ratio of pooled variances Var_A and Var_B}
}
CV values \code{CV_A} and \code{CV_B} can also be represented in percent.
To convert these to percentage values, one just multiply their raw results by 100%.
}
\examples{
\dontrun{
  library(data.table)
  data <- simulate_eqa_data2(list(n = 25, R = 3, cvx = 0.02, cvy = 0.3), AR = TRUE)
  data$SampleID <- as.character(data$SampleID)
  data$ReplicateID <- as.character(data$ReplicateID)
  print(global_prcision_estimates2(data = data) |> as.data.table())
}
}
