% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inside_deming2}
\alias{inside_deming2}
\title{Check Whether New Observations Are Inside Deming Prediction Intervals}
\usage{
inside_deming2(
  data,
  new_data,
  imprecision_estimates,
  R = 3L,
  R_ratio = 1,
  level = 0.99
)
}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain \code{MP_A} (response variable values)
and \code{MP_B} (predictor variable values).}

\item{new_data}{A \code{list} or \code{data.table}. Must contain \code{MP_A} (new response variable values)
and \code{MP_B} (new predictor variable values).}

\item{imprecision_estimates}{A \code{list} or \code{data.table}. Must contain \code{lambda}
and \code{Var_B}. See details for more information.}

\item{R}{A \code{integer} value. The number of replicated measurements. If the number of
replicated measurements is sample-dependent, use some summary statistic.
Defaults to \code{3}.}

\item{R_ratio}{A \code{double} value. If a different number of replicated measurements are used
in data than in new_data, specify the ratio here. Defaults to \code{1}.}

\item{level}{A \code{double} value. The desired nominal confidence level for
the estimated Deming prediction intervals. Must be between \code{0} and \code{1}.
Sensible values include values larger than or equal to \code{0.50}. Defaults to \code{0.99}.}
}
\value{
An \code{integer} vector of length equal to the number of rows in \code{new_data}.
}
\description{
Checks if \eqn{y_{n+j}} for \eqn{j = 1, \ldots, m} are inside Deming prediction intervals.
}
\details{
The Deming regression \eqn{\mathrm{level} \cdot 100\%} prediction intervals are calculated by

\eqn{PI_D(y_{n+j} | x_{n+j}) \approx b_0 + b_1 \cdot x_{n+j} \pm t_{n - 2, 1 - \alpha/2} \sqrt{\hat{\sigma}_{b_1}^2\Big[(\hat{x}_{n+j}^L - \hat{\mu})^2 + \tilde{\sigma}_h^2 \cdot R_{\mathrm{rat}}\Big] + R_{\mathrm{rat}} \cdot \Big(\frac{n+j}{n}\Big) \cdot (\tilde{\sigma}_v^2 + b_1^2 \cdot \tilde{\sigma}_h^2)}}

This function calculates \eqn{z_j = \mathbb{I}\big[y_{n+j} \in PI_D(y_{n+j} | x_{n+j})\big]}
for each \eqn{j = 1, \ldots, m}. Note that
\itemize{
  \item \eqn{b_1} is the Deming regression slope estimate of \eqn{\beta_1}.
  \item \eqn{\hat{\sigma}_{b_1}^2} is the estimated variance of the Deming regression slope estimator \eqn{b_1}.
  \item \eqn{\hat{x}_{n+j}^L} is an estimate of the latent \eqn{x_{n+j}^L}.
  \item \eqn{\hat{\mu}} is an estimate of the mean of \eqn{x_{i}^L}.
  \item \eqn{\tilde{\sigma}_h^2} is the method of moments estimator of the repeatability variance \eqn{\sigma_h^2}.
  \item \eqn{\tilde{\sigma}_v^2} is the method of moments estimator of the repeatability variance \eqn{\sigma_v^2}.
  \item \eqn{R_{\mathrm{rat}}} is the \code{R_ratio} value.
}

Missing values in \code{MP_A} or \code{MP_B} in \code{data} are allowed and will be silently removed
before \eqn{z_j} values are calculated. Missing values in \code{MP_A} or \code{MP_B} in \code{new_data}
are allowed, but corresponding \eqn{z_j} values will not be calculated in such cases.

Note that \code{imprecision_estimates} should contain the following variables:
\itemize{
  \item \code{lambda}: This an estimator of the ratio \eqn{\sigma_v^2 / \sigma_h^2}.
        How you estimate this ratio is up to you. However, a natural estimator is
        \eqn{\hat{\sigma}_v^2 / \hat{\sigma}_h^2}, where \eqn{\hat{\sigma}_v^2} and
        \eqn{\hat{\sigma}_h^2} are pooled variances based on replicated measurements.
  \item \code{Var_B}: This is an estimator of \eqn{\sigma_h^2}. How you estimate this
        ratio is up to you. However a natural estimator is \eqn{\hat{\sigma}_h^2},
        which is a pooled variance based on replicated measurements.
}
N.B., if either \code{lambda} or \code{Var_B} is missing or takes negative values
Ordinary least squares regression is used instead to calculate \eqn{z_j} values.
}
\examples{
\dontrun{
  print(1)
}
}
