% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{iterative_weights_estimation}
\alias{iterative_weights_estimation}
\title{Iterative Estimation of Smoothing Spline Weights}
\usage{
iterative_weights_estimation(
  x,
  y,
  u,
  interior_knots,
  df,
  lambda,
  df_max,
  tol = 0.25,
  window = 10,
  max_iter = 1
)
}
\description{
This function is used internally for estimating weights.
}
\keyword{internal}
