% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate Sample Excess Kurtosis of a Random Sample}
\usage{
kurtosis(x, na_rm = TRUE)
}
\arguments{
\item{x}{A \code{numeric} vector that is a random sample.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA}-values are
removed prior to calculation of sample excess kurtosis.}
}
\value{
A \code{double} that is the calculated sample excess kurtosis.
}
\description{
Calculate sample excess kurtosis based on a numeric vector \code{x}.
}
\details{
Calculates the sample excess kurtosis of a random sample \eqn{\lbrace x_i \rbrace_{i=1}^{N}}.
The sample excess kurtosis, attempts to estimate the theoretical excess kurtosis, \eqn{\Kappa}, by using
the following estimator

\eqn{\hat{\Kappa} = \frac{N(N+1)}{(N-1)(N-2)(N-3)} \cdot \frac{\sum_{i=1}^{N}(x_i - \overline{x})^4}{s^4} - 3 \cdot \frac{(N-1)^2}{(N-2)(N-3)}}.

Here, \eqn{s^2} is the unbiased sample variance.
Note that this estimator will be biased if not \eqn{x_i \sim \mathrm{N}(\mu, \sigma^2)}.
Note also that \eqn{\hat{\Kappa}} estimates the theoretical excess kurtosis and not the raw kurtosis.
Therefore, one should add \eqn{3} to the output to get the estimated raw kurtosis.
}
\examples{
\dontrun{
  y <- rnorm(n = 1000)
  kurt_y <- kurtosis(y)
  print(kurt_y)
}
}
