% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nfold_cv2}
\alias{nfold_cv2}
\title{N-fold cross-validation estimate of the mean squared prediction error of the smoothing spline model}
\arguments{
\item{sp}{A non-negative \code{double}. This is a scaled version of the raw smoothing parameter lambda and is usually what is optimized.}

\item{y}{A \code{numeric} vector sorted according to the order of the predictor \code{x}.}

\item{weights}{A \code{numeric} vector sorted according to the order of the predictor \code{x}.}

\item{B}{A B-spline basis \code{matrix} evaluated at the sorted and unique values of \code{x} .}

\item{BTB}{A \code{matrix} that is the matrix product between the transpose of \code{B} and \code{B}.}

\item{Omega}{The penalty \code{matrix} for the B-spline basis.}

\item{cross_validation}{A \code{logical} value. If set to \code{TRUE}, you are effectively prompting that the lambda inputs are equivalent to the spar smoothing parameter.}

\item{r}{The trace ratio of the matrices}
}
\value{
A \code{double} that signify the n-fold cross-validation score
}
\description{
N-fold cross-validation estimate of the mean squared prediction error of the smoothing spline model
}
\examples{
print(1)
}
