% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{numerical_gradient}
\alias{numerical_gradient}
\title{Approximate Numerical Gradient for a Function \code{f}}
\arguments{
\item{x}{A \code{double}. The point where \code{f} is to be evaluated.}

\item{f}{A \code{function}. If the function have more than 1 argument, ensure that these are fixed.}

\item{eps}{A \code{double}. The delta value in which the gradient are approximated.}

\item{r}{An \code{integer}. The number of Iterations utilized for the Richardson extrapolation algorithm.}
}
\value{
A \code{double} that signify the approximated gradient of \code{f} evaluated at \code{x}.
}
\description{
Numerical Gradient Approximation Using Richardson Extrapolation
}
\details{
Given a particular vale of \code{x}, approximates the gradient of \code{f(x)} at this point.
}
\examples{
example_fun <- function(x, a = 3){return((x - a)^2)}
gradient_at_minimum <- numerical_gradient(3, function(x) example_fun(3, 3))

# Expected to be approximately '0'
print(gradient_at_minimum)
}
