% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline_diagnostics.R
\name{optimal_df}
\alias{optimal_df}
\title{Internal Function for Estimating Optimal Values of DOF}
\usage{
optimal_df(data)
}
\description{
This function is used internally to readily get optimal values of \eqn{\mathrm{df}(c)}.
}
\keyword{internal}
