% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_matrix.R
\name{penalty_matrix}
\alias{penalty_matrix}
\title{O'Sullivan penalized splines penalty matrix}
\usage{
penalty_matrix(x_min, x_max, interior_knots)
}
\arguments{
\item{x_min}{A \code{double} indicating the lower boundary knot. The knot sequence for the B-spline basis will be augmented using this value.}

\item{x_max}{A \code{double} indicating the upper boundary knot. The knot sequence for the B-spline basis will be augmented using this value.}

\item{interior_knots}{A \code{numeric} vector containing the interior knots for the B-spline basis. Cannot include \code{x_min} or \code{x_max}.}
}
\value{
The O'Sullivan B-spline penalty matrix
}
\description{
O'Sullivan penalized splines penalty matrix
}
\examples{
print(1)
}
