% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{plot.smoothing_spline}
\alias{plot.smoothing_spline}
\title{Default Plot Method for Smoothing Spline Output}
\usage{
\method{plot}{smoothing_spline}(output, error = c("confidence", "prediction", "none"), level = 0.95)
}
\arguments{
\item{output}{A \code{smoothing_spline} object, which is the output of \code{smoothing_spline()} method.}

\item{error}{A \code{character} string value. Can be one of the following:
\itemize{
  \item \code{confidence}: Include estimated confidence bands around the smoothing spline.
  \item \code{prediction}: Include estimated prediction bands around the smoothing spline.
  \item \code{none}: Do not include estimated uncertainty bands around the smoothing spline.
}}

\item{level}{A \code{double}. The confidence level of the error interval.
Must be between \code{0} and \code{1}. Default value is \code{0.95}.}
}
\value{
Returns a \code{ggplot2} object. This plot contains the fitted \code{smoothing_spline} object,
        the data points used in fitting it, and uncertainty bands
        around the smoothing spline.
}
\description{
Default Plot Method for Smoothing Spline Output
}
\examples{
print(1)
}
