% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smoothing_spline.R
\name{predict_smoothing_spline}
\alias{predict_smoothing_spline}
\title{Prediction Interval Estimation for External Quality Assessment Data via Smoothing Splines}
\usage{
predict_smoothing_spline(
  data,
  new_data,
  df = NULL,
  lambda = NULL,
  df_max = 7.5,
  R_ratio = 1,
  level = 0.99,
  rounding = 3L
)
}
\arguments{
\item{data}{A \code{data.table}, \code{list} or \code{data.frame} object. Should include the ID column \code{SampleID} and the measurement columns \code{MP_A} and \code{MP_B}. Clinical sample measurements should be in here.}

\item{new_data}{A \code{data.table}, \code{list} or \code{data.frame} object. Can include the ID column \code{SampleID} and the measurement column \code{MP_A}, but the measurement column \code{MP_B} is mandatory. External quality assessment (EQA) material measurements should be in here.}

\item{df}{A optional \code{numeric} value greater than or equal to 2, but less than or equal to the number of clinical samples in \code{data}. If both \code{df} and \code{lambda} is set to \code{NULL}, n-fold cross-validation is used to obtain the optimal \code{df}.}

\item{lambda}{A optional \code{numeric} value greater than 0. If both \code{df} and \code{lambda} is set to \code{NULL}, n-fold cross-validation is used to obtain the optimal \code{lambda}.}

\item{df_max}{A \code{numeric} value between 2 and the number of clinical samples in \code{data}. If n-fold cross-validation is used to obtain the optimal effective number of degrees of freedom, what is its upper limit.}

\item{R_ratio}{A \code{numeric} value indicating the ratio of replicates between that number \code{data} and that number \code{new_data} are based on. Only relevant if the number of replicates of \code{data} and \code{new_data} differs.}

\item{level}{A \code{numeric} value representing the confidence level for the approximated prediction intervals. It should be between \code{0} and \code{1}. The default setting is \code{0.99}. Please adjust for simultaneous testing if pointwise prediction intervals are used for classifying more than one EQA material / reference material in the same IVD-MD comparison.}

\item{rounding}{An \code{integer} specifying the desired decimal places for the predictions and prediction intervals. The default setting is three, offering sufficient precision. The maximum limit is twelve due to the utilization of double numbers.}
}
\value{
A \code{data.table} object comprising approximated prediction interval data based on user inputs.
}
\description{
Prediction Interval Estimation for External Quality Assessment Data via Smoothing Splines
}
\details{
For optimal results, we recommended to include all \code{SampleID}, \code{MP_A}, and \code{MP_B} in \code{new_data} whenever possible. When constructing prediction bands, only \code{MP_B} should be available.
}
\examples{
print(1)
}
