% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smoothing_spline.R
\name{predict_smoothing_splines}
\alias{predict_smoothing_splines}
\title{Comparison-wise Prediction Interval Estimation for External Quality Assessment Data via Smoothing Splines}
\usage{
predict_smoothing_splines(
  data,
  new_data = NULL,
  weight_data = NULL,
  df = NULL,
  lambda = NULL,
  df_max = 7.5,
  R_ratio = 1,
  level = 0.99,
  simultaneous = FALSE,
  negative_ok = TRUE,
  rounding = 3L
)
}
\arguments{
\item{data}{A grouped \code{data.table}, \code{list} or \code{data.frame}
object by \code{comparison}. Should include the ID column \code{SampleID}
and the measurement columns \code{MP_A} and \code{MP_B}.
Clinical sample measurements should be in here.}

\item{new_data}{A grouped \code{data.table}, \code{list} or \code{data.frame}
object by \code{comparison}. Can include the ID column \code{SampleID}
and the measurement column \code{MP_A}, but the measurement column \code{MP_B} is mandatory.
External quality assessment (EQA) material measurements should be in here.}

\item{weight_data}{A grouped \code{data.table}, \code{list} or \code{data.frame}
object by \code{comparison}. Must include the ID column \code{SampleID}
and the weight columns \code{MP_A} and \code{MP_B}.
The computational weights are given by 1 / (\code{MP_A} + \code{MP_B}).
If set to \code{NULL}, no weights are used. See documentation of
\code{predict_smoothing_spline} for more information.}

\item{df}{A optional \code{numeric} vector of length equal to the number of
unique IVD-MD comparisons in \code{data}. It can alternatively be
a single \code{double} value or \code{NULL}, which then applies to all IVD-MD comparions.}

\item{lambda}{A optional \code{numeric} vector of length equal to the number of
unique IVD-MD comparisons in \code{data}. It can alternatively be
a single \code{double} value or \code{NULL}, which then applies to all IVD-MD comparions.}

\item{df_max}{A \code{numeric} value between 2 and the number of clinical samples in \code{data}.}

\item{R_ratio}{A \code{numeric} value indicating the ratio of replicates between
that number \code{data} and that number \code{new_data} are based on.
Only relevant if the number of replicates of \code{data} and \code{new_data} differs.}

\item{level}{A \code{numeric} value representing the confidence level for the
approximated prediction intervals. It should be between \code{0} and \code{1}.
The default setting is \code{0.99}.}

\item{simultaneous}{A \code{logical} value. If set to \code{TRUE},
prediction intervals are Bonferroni-corrected.
The default is set to \code{FALSE}.}

\item{negative_ok}{A \code{logical} value. If set to \code{TRUE}, negative values are allowed.
See details.}

\item{rounding}{An \code{integer} specifying the desired decimal places for the
predictions and prediction intervals. The default setting is \code{3L},
offering sufficient precision. The maximum limit is twelve
due to double precision.}
}
\value{
A \code{data.table} object comprising approximated prediction interval data based on user inputs.
}
\description{
Predict and estimate prediction intervals based on new observations using
smoothing splines grouped by \code{comparison}.
}
\details{
See documentation of \code{predict_smoothing_spline} for detailed information
on each of the function arguments.
}
\examples{
print(1)
}
