% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{print.smoothing_spline}
\alias{print.smoothing_spline}
\title{Default Print Method for Smoothing Spline Output}
\usage{
\method{print}{smoothing_spline}(output)
}
\arguments{
\item{output}{A \code{smoothing_spline} object, which is the output of \code{smoothing_spline()} method.}
}
\value{
Returns nothing, but prints compact an relevant information about the output of \code{smoothing_spline()}
}
\description{
Default Print Method for Smoothing Spline Output
}
\examples{
print(1)
}
