% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reconstruct_4_band_matrix}
\alias{reconstruct_4_band_matrix}
\title{Reconstruct a 4-banded matrix based on flattened data}
\usage{
reconstruct_4_band_matrix(x)
}
\arguments{
\item{x}{A \code{numeric} vector signifying the flattened version of the 4-banded matrix.}
}
\value{
A 4-banded \code{matrix}.
}
\description{
Reconstructs a 4-banded matrix based on flattened data.
}
\details{
Reconstruct a 4-banded matrix based on flattened data

This reconstruction assumes symmetrical 4-banded matrix. End-users should not bother with this function.
}
\examples{
\dontrun{
  print(1)
}
}
