% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_components.R
\name{replicate_simulate_components}
\alias{replicate_simulate_components}
\title{Replicate Simulation of Smoothing Spline Components}
\usage{
replicate_simulate_components(
  parameters,
  N = 100,
  shift = FALSE,
  attempt_fast = FALSE,
  include_pe = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{N}{An \code{integer} that represents the number of replicated component estimates.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A} and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the calculations are attempted speeded up by using some of the \code{smooth.spline()} functionality.}

\item{include_pe}{A \code{logical} value. If \code{TRUE}, predictions and prediction errors are returned.}

\item{parallel}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, simulations are done on multiple cores on your local machine for the duration of the function call.}

\item{max_cores}{An \code{integer} value between \code{2} and the number of cores you have on your computer. If you wish to use fewer cores than is on your computer, you should specify this here.}

\item{seed}{An \code{integer} value. If you wish to set a seed for reproducibility, set it here. If you do not want reproducibility, set this to \code{NULL}.}
}
\value{
A \code{data.table} object, grouped by the unique parameter combinations as well as component types
}
\description{
Replicate Simulation of Smoothing Spline Components
}
\examples{
print(1)
}
