% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_components.R
\name{replicate_simulate_components_internal}
\alias{replicate_simulate_components_internal}
\title{Internal Function for Component Simulation}
\usage{
replicate_simulate_components_internal(
  parameters,
  N = 100,
  shift = FALSE,
  attempt_fast = FALSE,
  include_pe = FALSE
)
}
\description{
This function repeats the simulate_components() function \code{N} times for a single set of simulation parameters.
}
\keyword{internal}
