% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_fun_of_samples}
\alias{resample_fun_of_samples}
\title{Resample cluster statistics based on EQA clinical sample data}
\arguments{
\item{data}{A \code{list} or a \code{data.table}. The mean-of-replicates
clinical sample data. Must contain \code{SampleID}, \code{MP_A} and \code{MP_B}.
The ID variable \code{SampleID} must be \code{character}.}
}
\value{
A \code{list} that contains the resampled \code{data}.
}
\description{
Resample cluster statistics based on EQA clinical sample data
}
\details{
This function is a very efficient method to resample aggregated sample statistics
based on clinical sample data.

It is convenient to combine this function with \code{smooth.commutability} functions such as
\itemize{
  \item \code{predict_smoothing_splines()} to estimate inside rates for a IVD-MD comparison
  \item \code{smoothing_spline()} to estimate bootstrap distribution of LOO-CV chosen effective degrees of freedom.
}
}
\examples{
library(data.table)
fictive_data <- simulate_eqa_data2(list(n = 25, R = 3, cvx = 0.01, cvy = 0.02))
resampled_data <- resample_fun_of_samples(fictive_data)
setDT(resampled_data)
print(resampled_data)
}
