% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resample_samples2}
\alias{resample_samples2}
\title{Resample clustered EQA clinical sample data}
\arguments{
\item{data}{A \code{list} or a \code{data.table}. Must contain \code{SampleID},
\code{ReplicateID}, \code{MP_A} and \code{MP_B}. The ID variables \code{SampleID}
and \code{ReplicateID} must be of character type for the function to operate correctly.}
}
\value{
A \code{list} containing the resampled EQA clinical sample data.
}
\description{
Resample clustered EQA clinical sample data
}
\details{
This function is a very efficient method to resample clinical sample data on sample-level.
It is convenient to combine this function with \code{fasteqa} functions such as
\itemize{
  \item \code{global_precision_estimates()} to estimate bootstrap imprecision confidence intervals
  \item \code{estimate_zeta()} to estimate bootstrap zeta confidence intervals
}
Alternatively it could be used to estimate \code{smooth.commutability} functions such as
\itemize{
  \item \code{estimate_zeta_ss()} to estimate bootstrap smoothing spline zeta confidence intervals
  \item \code{estimate_df()} to estimate bootstrap smoothing spline df confidence intervals
}
}
\examples{
library(data.table)
fictive_data <- simulate_eqa_data2(list(n = 25, R = 3, cvx = 0.01, cvy = 0.01), AR = TRUE)
resampled_data <- resample_samples2(fictive_data)
setDT(resampled_data)
print(resampled_data)
}
