% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_predictor_and_response.R
\name{shift_predictor_and_response}
\alias{shift_predictor_and_response}
\title{Internal Function for shifting roles of MP_A and MP_B}
\usage{
shift_predictor_and_response(data)
}
\description{
This function is used internally within \code{fix_one_predictor_and_response()}.
}
\keyword{internal}
