% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma_h_squared_deming2}
\alias{sigma_h_squared_deming2}
\title{Calculates the Deming Regression Estimate of \eqn{\sigma_h^2}.}
\usage{
sigma_h_squared_deming2(data, lambda = 1)
}
\arguments{
\item{data}{A \code{list} or \code{data.table}. Must contain \code{MP_A} (response variable values)
and \code{MP_B} (predictor variable values).}

\item{lambda}{A \code{double} value.}
}
\value{
An \code{double} that is the Deming regression estimate of \eqn{\sigma_h^2}.
}
\description{
Calculates the Deming regression estimate of \eqn{\sigma_h^2}.
}
\details{
This function is typically not relevant for end-users.
}
\examples{
\dontrun{
  print(1)
}
}
