% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_components.R
\name{simulate_components}
\alias{simulate_components}
\title{Simulate Smoothing Spline Components}
\usage{
simulate_components(
  parameters,
  shift = FALSE,
  attempt_fast = FALSE,
  include_pe = FALSE
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A} and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the smoothing spline fitting and predictions are attempted speeded up.}

\item{include_pe}{A \code{logical} value. If \code{TRUE}, predictions and prediction errors are returned.}
}
\value{
A named \code{integer} vector containing the relevant simulated components based on \code{parameters}
}
\description{
Simulate Smoothing Spline Components
}
\examples{
print(1)
}
