% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_components.R
\name{simulate_components_statistics}
\alias{simulate_components_statistics}
\title{Simulation of Statistics of The Smoothing Spline Components}
\usage{
simulate_components_statistics(
  parameters,
  N = 100,
  shift = FALSE,
  attempt_fast = FALSE,
  include_pe = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99,
  percent = TRUE,
  remove_extremes = FALSE
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{N}{An \code{integer} that represents the number of replicated component estimates.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A} and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the calculations are attempted speeded up by using some of the \code{smooth.spline()} functionality.}

\item{include_pe}{A \code{logical} value. If \code{TRUE}, predictions and prediction errors are returned.}

\item{parallel}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, simulations are done on multiple cores on your local machine for the duration of the function call.}

\item{max_cores}{An \code{integer} value between \code{2} and the number of cores you have on your computer. If you wish to use fewer cores than is on your computer, you should specify this here.}

\item{seed}{An \code{integer} value. If you wish to set a seed for reproducibility, set it here. If you do not want reproducibility, set this to \code{NULL}.}

\item{percent}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, all numbers that are possible to present as percentages are presented as percentages.}

\item{remove_extremes}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If \code{TRUE}, values \eqn{x_i} satisfying that \eqn{x_i < Q_2 - 10 \cdot \mathrm{IQR}\lbrace x_i \rbrace}
or \eqn{x_i > Q_3 + 10 \cdot \mathrm{IQR}\lbrace x_i \rbrace} are removed.}
}
\value{
A \code{data.table} object. Contains component summary statistics for each
        parameter value combination in \code{parameters} as well as each component.
        The following statistics are calculated for each combination:
        \itemize{
         \item \code{min}: The sample minimum
         \item \code{q1}: The empirical 1st percentile.
         \item \code{q2.5}: The empirical 2.5th percentile.
         \item \code{q5}: The empirical 5th percentile.
         \item \code{q10}: The empirical 10th percentile.
         \item \code{Q1}: The empirical first quartile of the values (25th percentile)
         \item \code{median}: The empirical second quartile of the values (50th percentile)
         \item \code{Q3}: The empirical third quartile of the values (75th percentile)
         \item \code{q0.90}: The empirical 90th percentile.
         \item \code{q0.95}: The empirical 95th percentile.
         \item \code{q0.975}: The empirical 97.5th percentile.
         \item \code{q0.99}: The empirical 99th percentile.
         \item \code{max}: The sample maximum
         \item \code{mean}: The sample mean
         \item \code{var}: The sample variance
         \item \code{sd}: The sample standard deviation
         \item \code{mad}: The sample median absolute deviation
         \item \code{skewness}: The sample skewness
         \item \code{kurtosis}: The sample excess kurtosis
        }
}
\description{
Simulation of Statistics of The Smoothing Spline Components
}
\examples{
print(1)
}
