% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_confidence_levels2.R
\name{simulate_confidence_levels2}
\alias{simulate_confidence_levels2}
\title{Simulate empirical confidence levels for a set of parameters}
\usage{
simulate_confidence_levels2(
  parameters,
  N = 100,
  m = 1,
  level = 0.95,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE,
  include_deming = FALSE,
  include_ci = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99,
  percent = TRUE
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{N}{An \code{integer} that represents the number of replicated inside checks.}

\item{m}{An \code{integer} larger than or equal to \code{1L}. How many external quality assessment materials should be simulated?}

\item{level}{A \code{double} that is larger than 0, but smaller than 1. Represents the confidence level of the estimated prediction intervals for the smoothing spline fit.}

\item{extrapolate_ok}{A \code{logical} value. If \code{TRUE}, test data can be simulated beyond the empirical range of the clinical sample data.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A} and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the predictions are attempted speeded up.}

\item{include_deming}{A \code{logical} value. If \code{TRUE}, the inside rates based on Deming regression are also included.}

\item{include_ci}{A \code{logical} value. If \code{TRUE}, approximate confidence intervals for are calculated.}

\item{parallel}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, simulations are done on multiple cores on your local machine for the duration of the function call.}

\item{max_cores}{An \code{integer} value between \code{2} and the number of cores you have on your computer. If you wish to use fewer cores than is on your computer, you should specify this here.}

\item{seed}{An \code{integer} value. If you wish to set a seed for reproducibility, set it here. If you do not want reproducibility, set this to \code{NULL}.}

\item{percent}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}). If set to \code{TRUE}, all numbers that are possible to present as percentages are presented as percentages.}
}
\value{
A \code{data.table}.
}
\description{
Simulates Monte Carlo approximated empirical confidence level, \eqn{\theta}
based on every unique combination of parameters found in \code{parameters},
based on \code{N} inside checks.
}
\details{
The approximate smoothing spline prediction interval for a new observation
\eqn{y_{n+1}} given the new predictor value \eqn{x_{n+1}} and a fitted smoothed
spline \eqn{\hat{f}} is defined to be:

\eqn{PI_T(y_{n+1} | x_{n+1}) \approx \hat{f}(x_{n+1}) \pm t_{n - \mathrm{df}(c), 1 - \alpha/2} \sqrt{\hat{\sigma} + \widehat{\mathrm{Var}}(\hat{f}(x_{n+1}))}}.

This function repeats the following steps \code{N} times for each unique parameter
combination in \code{parameters}:
\enumerate{
 \item Simulate \eqn{(x_i, y_i)} for \eqn{i = 1, 2, \ldots, n, n+1}
 \item Fit a smoothing spline using \eqn{(x_i, y_i)} for \eqn{i = 1, 2, \ldots, n}
 \item Calculate \eqn{z_j = \mathbb{I}(y_{n+1} \in PI_T(y_{n+1} | x_{n+1}))}
}
The empirical confidence level is calculated using \eqn{\hat{\theta} = \frac{\sum_{j=1}^{N}z_j}{N}}.
Assuming that \eqn{z_j \sim \mathrm{Bernoulli}(\theta)}, an asymptotically efficient estimator
for the variance of \eqn{\hat{\theta}} is

\eqn{\widehat{\mathrm{Var}}[\hat{\theta}] = \frac{\hat{\theta}(1-\hat{\theta})}{N}}

The corresponding standard error is denoted as \code{sem} in the output of this function.
An asymptotic \eqn{95\%} confidence interval for \eqn{\theta} is

\eqn{KI_\theta = \hat{\theta} \pm 1.96 \cdot \mathrm{sem}}

This asymptotic confidence interval is only calculated if \code{include_ci} is
set to \code{TRUE}.
}
\examples{
print(1)
}
