% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_eqa_data_custom_cpp}
\alias{simulate_eqa_data_custom_cpp}
\title{Simulate EQA data according to custom non-linear functions}
\arguments{
\item{parameters}{A \code{list} containing the simulation parameters.}

\item{type}{An \code{integer} value between 1 and 3. Which custom function should be simulated from.}

\item{AR}{A non-missing \code{logical} value. Set to \code{TRUE} if all replicated measurements should be returned. Otherwise the mean of replicates are returned}
}
\value{
A \code{list} containing the simulated EQA data for the specified \code{type}.
}
\description{
Simulates EQA data according to particular non-linear functions
}
\details{
Simulate EQA data according to custom non-linear functions

This function has no other use than in simulation studies.
}
\examples{
\dontrun{

  # Generate some fictional data
  simulate_eqa_data_custom_cpp(list(n = 25, R = 3, cil = 2, ciu = 10,
   cvx = 0.02, cvy = 0.03, cve = 0), type = 2, AR = FALSE)

}
}
