% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside.R
\name{simulate_inside_df}
\alias{simulate_inside_df}
\title{Internal Function for Inside Calculation Given That df is Given}
\usage{
simulate_inside_df(simulated_cs_data, simulated_eq_data, df, level = 0.95)
}
\description{
This function checks if external quality assessment materials are inside estimated prediction intervals given \code{df}
}
\keyword{internal}
