% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside.R
\name{simulate_inside_df_max}
\alias{simulate_inside_df_max}
\title{Internal Function for Inside Calculation Given That df_max is Given}
\usage{
simulate_inside_df_max(
  simulated_cs_data,
  simulated_eq_data,
  df_max,
  level = 0.95
)
}
\description{
This function checks if external quality assessment materials are inside estimated prediction intervals given \code{df_max}
}
\keyword{internal}
