% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside2.R
\name{simulate_insides2}
\alias{simulate_insides2}
\title{Simulate Clinical Sample Data and Evaluated Material Data and Check Whether
Inside PIs (Replicated)}
\usage{
simulate_insides2(
  parameters,
  N = 100,
  m = 1,
  level = 0.95,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE,
  include_deming = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame}.
The parameters and corresponding values to be used in
the simulation. Either \code{df} or \code{df_max}, must
be included. See \code{sim_eqa_data()} for other
parameters.}

\item{N}{An \code{integer}. The desired number of replicated inside checks.}

\item{m}{An \code{integer}. Must be larger than or equal to \code{1L}.
The desired number of evaluated materials to simulate. Defaults to
\code{1L}.}

\item{level}{A \code{double}. Must be between 0 and 1. The desired nominal
confidence level for the approximate prediction intervals.
Defaults to \code{0.95} (\eqn{95\%}).}

\item{extrapolate_ok}{A non-missing \code{logical} value. If \code{TRUE},
simulated evaluated material data are allowed to
produce measurements beyond the support of the
clinical samples IVD-MD measurements. Defaults to
\code{FALSE}.}

\item{shift}{A non-missing \code{logical} value. If \code{TRUE}, the roles
of \code{MP_A} and \code{MP_B} are allowed to shift. Can only
happen if \code{lambda < 0.5}. Defaults to \code{FALSE}.}

\item{attempt_fast}{A non-missing \code{logical} value. If \code{TRUE}, the
\code{smooth.spline()} is used to make predictions.}

\item{include_deming}{A non-missing \code{logical} value. If \code{TRUE},
the inside prediction interval indicator is also
calculated for \code{fg} Deming regression. Defaults
to \code{FALSE}.}

\item{parallel}{A non-missing \code{logical} value. If \code{TRUE},
simulations are done in parallel. May reduce runtime.}

\item{max_cores}{An \code{integer}. Must be between \code{2} and the number
of cores you have on your computer. If fewer cores than
the total number of cores on your computer is desired,
specify this here.}

\item{seed}{An \code{integer}. Ensuring reproducibility even if
\code{parallel = TRUE}. If reproducibility is unimportant, set
this to \code{NULL}.}
}
\value{
A \code{data.table}. The simulation results for the given \code{parameters}.
Expected to have \code{N} \eqn{\times} \code{m} rows. If \code{obs_tau} is
given in \code{parameters}, the resulting \code{data.table} is expected to
have \code{N} \eqn{\times} \code{m} \eqn{\times} \code{length(obs_tau)} rows.
}
\description{
Replicates \code{simulate_inside2()} \code{N} times.
}
\details{
See details in \code{?simulate_inside2()} for the finer details.

General note: Any function named \code{simulate_***}, is generally not
intended to be used by end-users. They are used in research and validity
testing.
}
\examples{
print(1)
}
