% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_inside2.R
\name{simulate_insides2_internal}
\alias{simulate_insides2_internal}
\title{Internal Function for Inside Calculation}
\usage{
simulate_insides2_internal(
  parameters,
  N = 100,
  m = 1,
  level = 0.95,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE,
  include_deming = FALSE
)
}
\description{
This function checks if external quality assessment materials are inside estimated prediction intervals
}
\keyword{internal}
