% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mspe.R
\name{simulate_pes}
\alias{simulate_pes}
\title{Replicate Calculation of Prediction Error for Each Parameter Combination}
\usage{
simulate_pes(
  parameters,
  N = 100,
  m = 1,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame}.
The parameters and corresponding values to be used in
the simulation. Either \code{df} or \code{df_max}, must
be included. See \code{sim_eqa_data()} for other
parameters.}

\item{N}{An \code{integer}. The desired number of replicated prediction error
calculations for each parameter combination in \code{parameters}.}

\item{m}{An \code{integer}. Must be larger than or equal to \code{1L}.
The desired number of evaluated materials to simulate. Defaults to
\code{1L}.}

\item{extrapolate_ok}{A non-missing \code{logical} value. If \code{TRUE},
simulated evaluated material data are allowed to
produce measurements beyond the support of the
clinical samples IVD-MD measurements. Defaults to
\code{FALSE}.}

\item{shift}{A non-missing \code{logical} value. If \code{TRUE}, the roles
of \code{MP_A} and \code{MP_B} are allowed to shift. Can only
happen if \code{lambda < 0.5}. Defaults to \code{FALSE}.}

\item{attempt_fast}{A non-missing \code{logical} value. If \code{TRUE}, the
\code{smooth.spline()} is used to make predictions.}

\item{parallel}{A non-missing \code{logical} value. If \code{TRUE},
simulations are done in parallel. May reduce runtime.}

\item{max_cores}{An \code{integer}. Must be between \code{2} and the number
of cores you have on your computer. If fewer cores than
the total number of cores on your computer is desired,
specify this here.}

\item{seed}{An \code{integer}. Ensuring reproducibility even if
\code{parallel = TRUE}. If reproducibility is unimportant, set
this to \code{NULL}.}
}
\value{
A \code{data.table}. The approximated mean squared prediction errors (MSPEs)
for each unique combination of parameters in \code{parameters}. The
resulting \code{data.table} is expected to have
\code{N} \eqn{\times} \code{m} \eqn{\times} \code{length(obs_tau)} rows.
}
\description{
Replicates \code{simulate_pe()} \code{N} times.
}
\details{
Additional details...
}
\examples{
# Required packages
library(fasteqa)

# Used parameters
parameters <- list(n = 25,
                   R = 3,
                   cvx = 0.01,
                   cvy = 0.01,
                   cil = 2,
                   ciu = 10,
                   obs_tau = 9,
                   type = 3,
                   df = 5)

# Simulated smoothing spline prediction errors
ss_pes <- simulate_pes(parameters = parameters,
                       N = 1000,
                       parallel = FALSE)

# Approximate prediction bias at y = 9
print(round(mean(ss_pes$pe, na.rm = TRUE), 4L))

# Approximate MSPE at y = 9
print(round(mean(ss_pes$pe^2, na.rm = TRUE), 4L))

}
