% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mspe.R
\name{simulate_pes_internal}
\alias{simulate_pes_internal}
\title{Internal Function for Prediction Error Calculation}
\usage{
simulate_pes_internal(
  parameters,
  N = 100,
  m = 1,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE
)
}
\description{
This function replicates prediction error calculation \code{N} times for
\code{parameters}
}
\keyword{internal}
