% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{skewness}
\alias{skewness}
\title{Calculate Sample Skewness of a Random Sample}
\usage{
skewness(x, na_rm = TRUE)
}
\arguments{
\item{x}{A \code{numeric} vector that is a random sample.}

\item{na_rm}{A \code{logical} value. If \code{TRUE}, \code{NA}-values are
removed prior to calculation of sample skewness.}
}
\value{
A \code{double} that is the calculated sample skewness.
}
\description{
Calculate sample skewness based on a numeric vector \code{x}.
}
\details{
Calculates the sample skewness of a random sample \eqn{\lbrace x_i \rbrace_{i=1}^{N}}.
The sample skewness, attempts to estimate the theoretical skewness, \eqn{\gamma}, by using
the following estimator

\eqn{\hat{\gamma} = \frac{\sqrt{N(N-1)}}{N-2} \cdot \frac{\frac{1}{N}\sum_{i=1}^{N}(x_i - \overline{x})^3}{\Big[\frac{1}{N}\sum_{i=1}^{N}(x_i - \overline{x})^2\Big]^{1.5}}}.

Note that this estimator will be biased if not \eqn{x_i \sim \mathrm{N}(\mu, \sigma^2)}.
}
\examples{
\dontrun{
  y <- rlnorm(n = 1000, meanlog = 0, sdlog = 0.25)
  skew_y <- skewness(y)
  print(skew_y)
}
}
