% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{smoothing_spline_residual_histogram}
\alias{smoothing_spline_residual_histogram}
\title{Internal Function for Plotting Smoothing Spline Residual Histograms}
\usage{
smoothing_spline_residual_histogram(plot_data, studentized = TRUE, df_t = 20)
}
\description{
This function is used internally within \code{smoothing_spline_residual_plots()}.
}
\keyword{internal}
