% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{smoothing_spline_residual_scatter_plot}
\alias{smoothing_spline_residual_scatter_plot}
\title{Internal Function for Plotting Smoothing Spline Residual Plots}
\usage{
smoothing_spline_residual_scatter_plot(
  plot_data,
  studentized = TRUE,
  df_t = 20
)
}
\description{
This function is used internally within \code{smoothing_spline_residual_plots()}.
}
\keyword{internal}
