% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{suggest_dfs}
\alias{suggest_dfs}
\title{Comparison-wise Suggestions for Optimal Starting Values for Effective Degrees of Freedom}
\usage{
suggest_dfs(
  data,
  ns_score = FALSE,
  use_weights = FALSE,
  output_type = c("mean", "median", "zd", "zm"),
  na_rm = TRUE
)
}
\arguments{
\item{data}{A \code{list} or \code{data.table} containing columns \code{comparison},
\code{SampleID}, \code{ReplicateID}, \code{MP_A} and \code{MP_B}.}

\item{ns_score}{A \code{logical}. If \code{TRUE}, calculates non-linearity scores for
each method used to derive optimal effective DOF values. See details.}

\item{use_weights}{A code{logical}. If \code{TRUE}, uses weights in the algorithms.
Weights are automatically determined based on \code{data}.}

\item{output_type}{A \code{character} value.}

\item{na_rm}{A \code{logical}. If \code{TRUE}, removes NA values before modeling. Note
that this affects algorithm results but not measurement variance or replicate
count calculations.}
}
\value{
A \code{data.table}
}
\description{
Comparison-wise Suggestions for Optimal Starting Values for Effective Degrees of Freedom
}
\examples{
print(1)
}
