% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_lambda.R
\name{to_lambda}
\alias{to_lambda}
\title{Convert the desired effective degrees of freedom to the computational smoothing parameter lambda}
\usage{
to_lambda(df, B, BTB, Omega)
}
\arguments{
\item{df}{A \code{double} that corresponds to the desired effective degrees of freedom for the smoothing spline model.}

\item{B}{A B-spline basis \code{matrix} evaluated at the sorted and unique values of the predictor values, \code{x}.}

\item{BTB}{A \code{matrix} that is the matrix product between the transpose of \code{B} and \code{B}.}

\item{Omega}{The O'Sullivan penalty \code{matrix} for the B-spline basis.}
}
\value{
A \code{double} that is the associated value of lambda to the provided \code{df}.
}
\description{
Convert the desired effective degrees of freedom to the computational smoothing parameter lambda
}
\examples{
print(1)
}
