% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeta_ss.R
\name{valdiate_estimate_zeta_ss}
\alias{valdiate_estimate_zeta_ss}
\title{Internal Function for validating entries}
\usage{
valdiate_estimate_zeta_ss(
  data,
  df,
  complex,
  advice,
  error = TRUE,
  funny = TRUE
)
}
\description{
This function is used internally validate entries in estimate_zeta_ss.
}
\keyword{internal}
