% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_smoothing_spline.R
\name{validate_prediction_data_components}
\alias{validate_prediction_data_components}
\title{Check input of predict_smoothing_spline}
\usage{
validate_prediction_data_components(data, new_data, weight_data)
}
\description{
This function is used internally for error handling regarding bad input.
}
\keyword{internal}
