% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{weight_function}
\alias{weight_function}
\title{Calculate Smooth Weights}
\usage{
weight_function(
  data,
  weight_data,
  impr_data,
  lower = 0.25,
  df = 5,
  inv = FALSE,
  output_type = c("vector", "table"),
  plot_weight_function = FALSE
)
}
\description{
This function is used internally for calculating smoother weights.
}
\keyword{internal}
