% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{zeta_deriv_df_plot}
\alias{zeta_deriv_df_plot}
\title{Internal Function for Plotting Zeta Derivative Loss With Respect To Effective DOF}
\usage{
zeta_deriv_df_plot(
  plot_data,
  which_plot = c("angle", "zeta_deriv_loss"),
  df_unit = FALSE
)
}
\description{
This function is used internally within \code{zeta_df_plots()} method.
}
\keyword{internal}
