% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{zeta_df_plot}
\alias{zeta_df_plot}
\title{Internal Function for Plotting Zeta Loss With Respect To Effective DOF}
\usage{
zeta_df_plot(
  plot_data,
  which_plot = c("zeta", "zeta_loss"),
  df_unit = FALSE,
  area = NULL
)
}
\description{
This function is used internally within \code{zeta_df_plots()} method.
}
\keyword{internal}
