% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{zeta_df_plots}
\alias{zeta_df_plots}
\title{Internal Function for Plotting Zeta Derivative Loss With Respect To Effective DOF}
\usage{
zeta_df_plots(
  df = NULL,
  x,
  y,
  weights,
  var_v,
  var_h,
  all_knots,
  loss = "abs",
  R = 3,
  mor = TRUE,
  targets = c(1, 0),
  which_plot = c("all", "zeta", "zeta_loss", "angle", "zeta_deriv_loss"),
  df_unit = FALSE
)
}
\description{
This function is used internally within \code{suggest_df()} method.
}
\keyword{internal}
