% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_df.R
\name{zeta_loss}
\alias{zeta_loss}
\title{Internal Function for Calculting Zeta Loss With Respect To Effective DOF}
\usage{
zeta_loss(
  df,
  x,
  y,
  weights,
  var_v,
  var_h,
  all_knots,
  loss = "abs",
  R = 3,
  mor = TRUE,
  target = 1
)
}
\description{
This function is used internally within multiple functions.
}
\keyword{internal}
