% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{fast_smoothing_spline}
\alias{fast_smoothing_spline}
\title{Internal Function for Faster Smoothing Spline Fitting}
\usage{
fast_smoothing_spline(
  orig_x,
  y,
  x,
  weights,
  all_knots,
  df = NULL,
  lambda = NULL,
  df_max = 7.5
)
}
\description{
This function is used internally for fitting a smoothing spline faster.
}
\keyword{internal}
