% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mspe.R
\name{simulate_mspe}
\alias{simulate_mspe}
\title{Simulate Mean Squared Prediction Error for Smoothing Spline}
\usage{
simulate_mspe(
  parameters,
  N = 100,
  m = 1,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE,
  parallel = TRUE,
  max_cores = 4,
  seed = 99,
  percent = TRUE,
  log_mspe = FALSE,
  include_ici = FALSE,
  trim = FALSE
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{N}{An \code{integer} that represents the number of replicated MSPE calculations
based on each parameter combination.}

\item{m}{An \code{integer} larger than or equal to \code{1L}.
How many external quality assessment materials should be simulated?}

\item{extrapolate_ok}{A \code{logical} value. If \code{TRUE}, test data can be
simulated beyond the empirical range of the clinical sample data.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A}
and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the predictions are attempted speeded up.}

\item{parallel}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If set to \code{TRUE}, simulations are done on multiple cores on
your local machine for the duration of the function call.}

\item{max_cores}{An \code{integer} value between \code{2} and the number of cores
you have on your computer. If you wish to use fewer cores than
is on your computer, you should specify this here.}

\item{seed}{An \code{integer} value. If you wish to set a seed for reproducibility, set it here.
If you do not want reproducibility, set this to \code{NULL}.}

\item{percent}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If set to \code{TRUE}, all numbers that are possible to present
as percentages are presented as percentages.}

\item{log_mspe}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If set to \code{TRUE}, simulated MSPE values are log-transformed.
This may be useful for plotting.}

\item{include_ici}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If set to \code{TRUE}, first and third quartiles of simulated
MSPE values are returned part of the output object.}

\item{trim}{A non-missing \code{logical} value (\code{TRUE} / \code{FALSE}).
If set to \code{TRUE}, trimmed mean with fraction \eqn{0.5\%} is used.}
}
\value{
A \code{data.table}.
}
\description{
Simulate Mean Squared Prediction Error for Smoothing Spline
}
\examples{
print(1)
}
