% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mspe.R
\name{simulate_pe}
\alias{simulate_pe}
\title{Simulate Prediction Errors}
\usage{
simulate_pe(
  parameters,
  m = 1,
  extrapolate_ok = FALSE,
  shift = FALSE,
  attempt_fast = FALSE
)
}
\arguments{
\item{parameters}{A \code{list}, \code{data.table} or \code{data.frame} with column names representing the parameters to be included in the simulations.
Must in general include \code{n}, \code{R}, \code{cvx}, \code{cvy}, and \code{df} or \code{df_max}. If \code{type} is included, data will
be simulated from custom non-linear functions. Note that \code{cil} and \code{ciu} also must be included if \code{type} is included.}

\item{m}{A \code{integer} larger than or equal to \code{1L}. How many external quality assessment materials should be simulated?}

\item{extrapolate_ok}{A \code{logical} value. If \code{TRUE}, test data can be simulated beyond the empirical range of the clinical sample data.}

\item{shift}{A \code{logical} value. If \code{TRUE}, the roles of \code{MP_A} and \code{MP_B} shift if \code{lambda < 0.5}.}

\item{attempt_fast}{A \code{logical} value. If \code{TRUE}, the predictions are attempted speeded up.}
}
\value{
An \code{numeric} vector of length \code{m}.
}
\description{
Simulate Prediction Errors
}
\examples{
print(1)
}
